fix : nav 수정 및 creditApp 모달창 수정
