+++ src/pages/ListPage.js
@@ -3,7 +3,6 @@ import CreditApp from '../components/listpage/CreditApp';
 const ListPage = () => {
   return (
     <div>
-      <h1>List 페이지</h1>
       <CreditApp />
     </div>
   );
