+++ src/pages/Main.js
@@ -5,7 +5,6 @@ const Main = () => {
   return (
     <>
       <Nav />
-
       <Outlet />
     </>
   );
