+++ src/styles/Nav.css
@@ -1,0 +1,24 @@
+.nav__section {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  position: relative;
+  height: 80px;
+  max-width: 1920px;
+  gap: 0px;
+  margin: 0 auto;
+}
+
+.nav__logo {
+  cursor: pointer;
+  height: 32px;
+}
+
+.nav__profile {
+  cursor: pointer;
+  position: absolute;
+  right: 20px;
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+}
